backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
openai
motor
qdrant-client
python-multipart
firebase-admin
python-dotenv
requests
tqdm
